+++ src/main/java/github/kaloyanov5/merkantil/dto/response/OrderResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,9 +15,9 @@ public class OrderResponse {
     private String symbol;
     private String side;
     private Integer quantity;
-    private Double executedPrice;
-    private Double limitPrice;
-    private Double totalAmount;
+    private BigDecimal executedPrice;
+    private BigDecimal limitPrice;
+    private BigDecimal totalAmount;
     private String orderType;
     private String status;
     private LocalDateTime timestamp;
