+++ src/main/java/github/kaloyanov5/merkantil/dto/response/StockHistoryResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -11,9 +12,9 @@ import java.time.LocalDate;
 @NoArgsConstructor
 public class StockHistoryResponse {
     private LocalDate date;
-    private Double open;
-    private Double high;
-    private Double low;
-    private Double close;
+    private BigDecimal open;
+    private BigDecimal high;
+    private BigDecimal low;
+    private BigDecimal close;
     private Long volume;
 }
